+++ lib/src/model/project_context.dart
@@ -1,0 +1,15 @@
+class ProjectContext {
+  const ProjectContext({
+    required this.rootPath,
+    required this.packageName,
+    required this.dartFiles,
+    required this.pubspecPath,
+  });
+
+  final String rootPath;
+  final String packageName;
+  final String pubspecPath;
+  final List<String> dartFiles;
+
+  int get dartFileCount => dartFiles.length;
+}